Refactored the hook to initialise the driver to use the IObjectContainer, this allows the implicit passing of the driver to specstep files without the need for an Object Repository. This is for context injection purposes.
