+++ UtilityClasses/IWebDriverExtensions.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using OpenQA.Selenium;
+﻿using OpenQA.Selenium;
 using TTCBDD.Configuration;
 
 namespace TTCBDD.UtilityClasses
